Gtk: Refactor
This is mainly separation of classes into their own files and a few class renames,
so we have better readability and understanding of symbol interdependiencies.
